+++ FRED/src/.dockerignore
