+++ FRED/src/requirements.txt
@@ -1,0 +1,7 @@
+Selenium
+Webdriver (ChromeDriver)
+Pandas
+Pathlib
+Website used = https://fred.stlouisfed.org/categories/22.
+Scheduler 
+Time
