+++ data/Interest rates for 48 month loan.csv
@@ -1,0 +1,200 @@
+,DATE,Finance rate on consumer installment loans 48 Month Loan
+0,1972-02-01,10.20
+1,1972-05-01,9.96
+2,1972-08-01,10.02
+3,1972-11-01,10.02
+4,1973-02-01,10.05
+5,1973-05-01,10.05
+6,1973-08-01,10.25
+7,1973-11-01,10.49
+8,1974-02-01,10.53
+9,1974-05-01,10.63
+10,1974-08-01,11.15
+11,1974-11-01,11.57
+12,1975-02-01,11.51
+13,1975-05-01,11.39
+14,1975-08-01,11.31
+15,1975-11-01,11.24
+16,1976-02-01,11.18
+17,1976-05-01,11.01
+18,1976-08-01,11.07
+19,1976-11-01,11.02
+20,1977-02-01,11.14
+21,1977-05-01,10.81
+22,1977-08-01,10.86
+23,1977-11-01,10.86
+24,1978-02-01,10.86
+25,1978-05-01,10.84
+26,1978-08-01,11.09
+27,1978-11-01,11.29
+28,1979-02-01,11.60
+29,1979-05-01,11.73
+30,1979-08-01,11.88
+31,1979-11-01,12.85
+32,1980-02-01,13.28
+33,1980-05-01,15.72
+34,1980-08-01,13.91
+35,1980-11-01,14.29
+36,1981-02-01,15.84
+37,1981-05-01,16.04
+38,1981-08-01,16.92
+39,1981-11-01,17.36
+40,1982-02-01,17.05
+41,1982-05-01,17.20
+42,1982-08-01,17.08
+43,1982-11-01,15.97
+44,1983-02-01,14.81
+45,1983-05-01,13.90
+46,1983-08-01,13.50
+47,1983-11-01,13.46
+48,1984-02-01,13.32
+49,1984-05-01,13.53
+50,1984-08-01,14.08
+51,1984-11-01,13.91
+52,1985-02-01,13.37
+53,1985-05-01,13.16
+54,1985-08-01,12.72
+55,1985-11-01,12.39
+56,1986-02-01,12.29
+57,1986-05-01,11.45
+58,1986-08-01,11.00
+59,1986-11-01,10.58
+60,1987-02-01,10.35
+61,1987-05-01,10.23
+62,1987-08-01,10.37
+63,1987-11-01,10.86
+64,1988-02-01,10.72
+65,1988-05-01,10.55
+66,1988-08-01,10.93
+67,1988-11-01,11.22
+68,1989-02-01,11.76
+69,1989-05-01,12.44
+70,1989-08-01,12.13
+71,1989-11-01,11.94
+72,1990-02-01,11.80
+73,1990-05-01,11.82
+74,1990-08-01,11.89
+75,1990-11-01,11.62
+76,1991-02-01,11.60
+77,1991-05-01,11.28
+78,1991-08-01,11.06
+79,1991-11-01,10.61
+80,1992-02-01,9.89
+81,1992-05-01,9.52
+82,1992-08-01,9.15
+83,1992-11-01,8.60
+84,1993-02-01,8.57
+85,1993-05-01,8.17
+86,1993-08-01,7.98
+87,1993-11-01,7.63
+88,1994-02-01,7.54
+89,1994-05-01,7.76
+90,1994-08-01,8.41
+91,1994-11-01,8.75
+92,1995-02-01,9.70
+93,1995-05-01,9.78
+94,1995-08-01,9.44
+95,1995-11-01,9.36
+96,1996-02-01,9.12
+97,1996-05-01,8.93
+98,1996-08-01,9.11
+99,1996-11-01,9.03
+100,1997-02-01,8.92
+101,1997-05-01,9.20
+102,1997-08-01,8.99
+103,1997-11-01,8.96
+104,1998-02-01,8.89
+105,1998-05-01,8.69
+106,1998-08-01,8.70
+107,1998-11-01,8.62
+108,1999-02-01,8.34
+109,1999-05-01,8.31
+110,1999-08-01,8.44
+111,1999-11-01,8.67
+112,2000-02-01,8.88
+113,2000-05-01,9.22
+114,2000-08-01,9.62
+115,2000-11-01,9.64
+116,2001-02-01,9.17
+117,2001-05-01,8.67
+118,2001-08-01,8.30
+119,2001-11-01,7.86
+120,2002-02-01,7.80
+121,2002-05-01,7.74
+122,2002-08-01,7.58
+123,2002-11-01,7.34
+124,2003-02-01,7.11
+125,2003-05-01,7.05
+126,2003-08-01,6.76
+127,2003-11-01,6.82
+128,2004-02-01,6.72
+129,2004-05-01,6.43
+130,2004-08-01,6.55
+131,2004-11-01,6.71
+132,2005-02-01,6.86
+133,2005-05-01,6.92
+134,2005-08-01,7.08
+135,2005-11-01,7.43
+136,2006-02-01,7.39
+137,2006-05-01,7.60
+138,2006-08-01,7.95
+139,2006-11-01,7.92
+140,2007-02-01,7.74
+141,2007-05-01,7.92
+142,2007-08-01,7.82
+143,2007-11-01,7.59
+144,2008-02-01,7.27
+145,2008-05-01,6.84
+146,2008-08-01,6.92
+147,2008-11-01,7.06
+148,2009-02-01,6.92
+149,2009-05-01,6.79
+150,2009-08-01,6.61
+151,2009-11-01,6.55
+152,2010-02-01,6.45
+153,2010-05-01,6.26
+154,2010-08-01,6.24
+155,2010-11-01,5.87
+156,2011-02-01,5.85
+157,2011-05-01,5.79
+158,2011-08-01,5.89
+159,2011-11-01,5.40
+160,2012-02-01,5.07
+161,2012-05-01,4.88
+162,2012-08-01,4.88
+163,2012-11-01,4.82
+164,2013-02-01,4.69
+165,2013-05-01,4.13
+166,2013-08-01,4.46
+167,2013-11-01,4.42
+168,2014-02-01,4.23
+169,2014-05-01,4.50
+170,2014-08-01,4.16
+171,2014-11-01,4.06
+172,2015-02-01,4.53
+173,2015-05-01,4.15
+174,2015-08-01,4.09
+175,2015-11-01,4.00
+176,2016-02-01,4.17
+177,2016-05-01,4.33
+178,2016-08-01,4.25
+179,2016-11-01,4.45
+180,2017-02-01,4.52
+181,2017-05-01,4.67
+182,2017-08-01,4.42
+183,2017-11-01,4.81
+184,2018-02-01,4.74
+185,2018-05-01,5.05
+186,2018-08-01,5.05
+187,2018-11-01,5.29
+188,2019-02-01,5.5
+189,2019-05-01,5.35
+190,2019-08-01,5.27
+191,2019-11-01,5.45
+192,2020-02-01,5.29
+193,2020-05-01,5.13
+194,2020-08-01,4.98
+195,2020-11-01,4.95
+196,2021-02-01,5.21
+197,2021-05-01,5.28
+198,2021-08-01,5.14
