+++ data/Interest rates for 60 month loan.csv
@@ -1,0 +1,62 @@
+,DATE,Finance rate on consumer installment loans 60 Month Loan
+0,2006-08-01,7.82
+1,2006-11-01,7.78
+2,2007-02-01,7.71
+3,2007-05-01,7.78
+4,2007-08-01,7.77
+5,2007-11-01,7.60
+6,2008-02-01,7.18
+7,2008-05-01,6.93
+8,2008-08-01,6.94
+9,2008-11-01,7.12
+10,2009-02-01,6.96
+11,2009-05-01,6.93
+12,2009-08-01,6.71
+13,2009-11-01,6.59
+14,2010-02-01,6.45
+15,2010-05-01,6.38
+16,2010-08-01,6.23
+17,2010-11-01,6.05
+18,2011-02-01,6.10
+19,2011-05-01,5.83
+20,2011-08-01,5.62
+21,2011-11-01,5.25
+22,2012-02-01,5.02
+23,2012-05-01,4.87
+24,2012-08-01,4.75
+25,2012-11-01,4.64
+26,2013-02-01,4.60
+27,2013-05-01,4.34
+28,2013-08-01,4.38
+29,2013-11-01,4.52
+30,2014-02-01,4.37
+31,2014-05-01,4.46
+32,2014-08-01,4.11
+33,2014-11-01,4.07
+34,2015-02-01,4.27
+35,2015-05-01,4.21
+36,2015-08-01,4.25
+37,2015-11-01,4.05
+38,2016-02-01,4.11
+39,2016-05-01,4.15
+40,2016-08-01,4.25
+41,2016-11-01,4.05
+42,2017-02-01,4.21
+43,2017-05-01,4.24
+44,2017-08-01,4.36
+45,2017-11-01,4.51
+46,2018-02-01,4.75
+47,2018-05-01,4.99
+48,2018-08-01,4.96
+49,2018-11-01,5.36
+50,2019-02-01,5.24
+51,2019-05-01,5.36
+52,2019-08-01,5.27
+53,2019-11-01,5.37
+54,2020-02-01,5.15
+55,2020-05-01,5.14
+56,2020-08-01,4.98
+57,2020-11-01,4.8
+58,2021-02-01,4.96
+59,2021-05-01,5.05
+60,2021-08-01,4.6
