--- dummy-data-product/src/Dockerfile
@@ -1,7 +1,0 @@
-FROM python:3.8-slim-buster
-
-COPY . .
-
-RUN pip3 install -r requirements.txt
-
-ENTRYPOINT ["python3"]
