--- dummy-data-product/src/requirements.txt
@@ -1,7 +1,0 @@
-Selenium
-Webdriver (ChromeDriver)
-Pandas
-Pathlib
-Website used = https://fred.stlouisfed.org/categories/22.
-Scheduler 
-Time
